Rework settings and add fallback functionality
Most of the settings have been reorganized. None of them work yet, but
that's next on the agenda.

As for the fallback functionality: now, if morph.io isn't working (for
whatever reason), then nasa.gov will be used in its place. Also, for
image downloads, if the full image doesn't exist, then the thumbnail
will be downloaded instead.